requirements.txt
fastapi==0.115.6
pydantic==2.10.3
python_bcrypt==0.3.2
